2015 - day 20 - part 2 - with help from reddit
I found a more performant way to generate the sequence
of divisors on-demand using sequence generators and
that improved performance of the part 1 calculations
by an order of magnitude (from 18m -> 7s) but my
part 2 calculations were consistently giving me the
wrong answer (1006920) which was too high, meaning
that they were calculating something wrong.

finally I looked up a solution on reddit and found
an F# algorithm which was actually quite close to
the way I was approaching it but this one gave
me the correct answer.